Refactor SQLite storage module by introducing modular structure
This commit reorganizes the SQLite storage module by splitting the functionality into multiple files: `current_state.rs`, `filesystem.rs`, `introspection.rs`, `schema.rs`, and `tests.rs`. The `SqliteConnectionPool` struct is updated to include methods for connecting to the database and executing SQL commands. Additionally, migration functions for handling branch-scoped primary keys in artefacts and edges are added, enhancing the schema management capabilities. The changes improve code maintainability and testability while ensuring existing functionalities remain intact.
